Give integration tests their own module
