+++ cosmos-integration-tests/src/main/resources/default-repositories.json
