+++ cosmos-integration-tests/src/main/resources/logback-test.xml
