+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/PackageAddSpec.scala
@@ -1,6 +1,7 @@
 package com.mesosphere.cosmos
 
 import _root_.io.circe.Json
+import com.mesosphere.TestUtil.eventualFuture
 import com.mesosphere.cosmos.circe.Decoders.decode
 import com.mesosphere.cosmos.converter.Response._
 import com.mesosphere.cosmos.http.CosmosRequests
@@ -15,7 +16,6 @@ import com.mesosphere.cosmos.storage.StagedPackageStorage
 import com.mesosphere.cosmos.test.CosmosIntegrationTestClient.CosmosClient
 import com.mesosphere.cosmos.test.CosmosIntegrationTestClient.PackageStorageClient
 import com.mesosphere.cosmos.test.InstallQueueFixture
-import com.mesosphere.cosmos.test.TestUtil
 import com.mesosphere.universe
 import com.mesosphere.universe.test.TestingPackages
 import com.mesosphere.universe.v3.syntax.PackageDefinitionOps._
@@ -152,7 +152,7 @@ final class PackageAddSpec
     assertSamePackage(
       expectedSupportedPackageDefinition,
       Await.result(
-        TestUtil.eventualFuture(
+        eventualFuture(
           () => packageStorage.readPackageDefinition(
             expectedSupportedPackageDefinition.packageCoordinate
           )
