+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/ServiceStartSpec.scala
