+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/finagle/FinagleStreamingSpec.scala
