+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/handler/CapabilitiesHandlerSpec.scala
