+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/http/CosmosRequests.scala
