+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/repository/PackageRepositorySpec.scala
