+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/repository/UniverseClientSpec.scala
