+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/rpc/v1/model/ErrorResponseSpec.scala
