+++ cosmos-integration-tests/src/main/scala/com/mesosphere/cosmos/test/CosmosIntegrationTestClient.scala
