+++ cosmos-test-util/src/main/scala/com/mesosphere/TestUtil.scala
@@ -1,0 +1,16 @@
+package com.mesosphere
+
+import com.twitter.util.Future
+
+object TestUtil {
+
+  def eventualFuture[T](
+    future: () => Future[Option[T]]
+  ): Future[T] = {
+    future().flatMap {
+      case Some(value) => Future.value(value)
+      case None => eventualFuture(future)
+    }
+  }
+
+}
