+++ cosmos-test-util/src/main/scala/com/mesosphere/cosmos/http/HttpRequest.scala
